make sure grabber return np.float32 not np.float64

+++ pingu/corpus/grabber/signal.py
@@ -63,7 +63,7 @@ class FramedSignalGrabber(object):
         frame = np.array(range[4][sample_start:sample_end])
         max_value = np.iinfo(frame.dtype).max
 
-        frame = frame.astype(np.float) / max_value
+        frame = frame.astype(np.float32) / max_value
 
         if sample_pad:
             frame = np.pad(frame, (0, sample_pad), mode='constant', constant_values=0)
@@ -184,7 +184,7 @@ class FramedSignalGrabber(object):
 
         ('a', 'b', 'd') --> [1,1,0,1]
         """
-        vec = np.zeros(len(self.labels)).astype(np.float)
+        vec = np.zeros(len(self.labels)).astype(np.float32)
 
         for label in labels:
             index = self.labels.index(label.value)
